refactor: create categoryData utils
